+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Fixed
+
+- Fix an issue where the preprocessor would remove comments located immediately
+  after the token that ends the replacement list.
+
+
 ## [v0.2.1] - 2020-08-30
 
 ### Fixed